:green_heart: [update] DBアクセス処理を見直し
■ DBアクセス処理を見直した。
  ・ クエリを修正
  ・ 結果が空(NULL)の場合の対策を追加した。

+++ uninstall.php
@@ -16,10 +16,17 @@ function nt_wpcf7sn_delete_plugin() {
 
 	// コンタクトフォームオプション削除
 	$options = $wpdb->get_results( "
-		SELECT * FROM $wpdb->options
-		WHERE option_name like 'nt_wpcf7sn_%'
+		SELECT *
+		  FROM $wpdb->options
+		WHERE 1 = 1
+		  option_name like 'nt_wpcf7sn_%'
+		ORDER BY option_name
 	" );
-	
+
+	if ( empty( $options ) ) {
+		return;
+	}
+
 	foreach ( $options as $option ) {
 		delete_option( $option->option_name );
 	}
